# Conflicts:
#	calculator/pom.xml
#	calculator/src/main/java/com/market/banica/calculator/service/CalculatorServiceImpl.java
#	calculator/src/main/java/com/market/banica/calculator/service/contract/CalculatorService.java
#	market-generator/pom.xml
#	market-generator/src/main/java/com/market/banica/generator/service/MarketService.java
#	market-generator/src/main/java/com/market/banica/generator/service/MarketServiceImpl.java
#	persister/pom.xml
#	pom.xml

+++ calculator/src/main/java/com/market/banica/calculator/service/contract/CalculatorService.java
@@ -15,5 +15,5 @@ public interface CalculatorService {
      * @param quantity quantity of the item
      * @return recipe for item
      */
-    RecipeDTO getRecipe(String itemName, double quantity);
+    RecipeDTO getRecipe(String itemName, int quantity);
 }
